criando popup de delete em projetos

+++ src/app/pages/projeto/listagem-projeto/listagem-projeto.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { FormGroup, FormBuilder } from '@angular/forms';
 import { Router } from '@angular/router';
-import { MessageService } from 'primeng/api';
+import { ConfirmationService, MessageService } from 'primeng/api';
 import { ClienteService } from 'src/app/services/cliente-service/cliente.service';
 import { ProjetoService } from 'src/app/services/projeto/projeto.service';
 import { Projeto, ProjetoListagem } from 'src/app/models/projeto/projeto';
@@ -14,7 +14,7 @@ import { StatusService } from 'src/app/services/status/status.service';
   selector: 'app-listagem-projeto',
   templateUrl: './listagem-projeto.component.html',
   styleUrl: './listagem-projeto.component.css',
-  providers: [MessageService]
+  providers: [MessageService, ConfirmationService]
 })
 export class ListagemProjetoComponent {
   form!: FormGroup;
@@ -24,7 +24,7 @@ export class ListagemProjetoComponent {
   clientes: Cliente[] = [];
   status: Status[] = [];
 
-  constructor(private fb: FormBuilder, private clienteService: ClienteService, private statusService: StatusService, private service: ProjetoService, private messageService: MessageService, private router: Router) { }
+  constructor(private fb: FormBuilder, private clienteService: ClienteService, private statusService: StatusService, private service: ProjetoService, private messageService: MessageService, private router: Router, private confirmationService: ConfirmationService) { }
 
   show(type: string, title: string, message: string) {
     this.messageService.add({ severity: type, summary: title, detail: message });
@@ -101,26 +101,41 @@ export class ListagemProjetoComponent {
     }
   }
 
-  deletar(id: number) {
+  deletar(event: Event, id: string) {
     if (id) {
       this.loading = true;
-      this.service.delete(id).subscribe({
-        next: (response: any) => {
-          if (response) {
-            this.show('success', 'Excluir Projeto', 'Projeto excluido com sucesso!');
-            this.getProjetos();
-            this.loading = false;
-          } else {
-            this.show('error', 'Excluir Projeto', 'Não foi possível excluir o projeto, tente novamente mais tarde');
-            this.loading = false;
-          }
+      this.confirmationService.confirm({
+        target: event.target as EventTarget,
+        message: 'Você tem certeza que deseja excluir este projeto?',
+        header: 'Excluir Projeto',
+        icon: 'pi pi-exclamation-triangle',
+        acceptIcon: "none",
+        rejectIcon: "none",
+        acceptButtonStyleClass: "p-button-danger p-button-text",
+        rejectButtonStyleClass: "p-button-text p-button-text",
+        accept: () => {
+          this.service.delete(id).subscribe({
+            next: (response: any) => {
+              if (response) {
+                this.show('success', 'Excluir Projeto', 'Projeto excluido com sucesso!');
+                this.getProjetos();
+                this.loading = false;
+              } else {
+                this.show('error', 'Excluir Projeto', 'Não foi possível excluir o projeto, tente novamente mais tarde');
+                this.loading = false;
+              }
 
+            },
+            error: (error: any) => {
+              this.show('error', 'Excluir Projeto', `${error.error.error ? error.error.error : 'Não foi possível excluir o projeto, tente novamente mais tarde'}`);
+              this.loading = false;
+            }
+          })
         },
-        error: (error: any) => {
-          this.show('error', 'Excluir Projeto', `${error.error.error ? error.error.error : 'Não foi possível excluir o projeto, tente novamente mais tarde'}`);
+        reject: () => {
           this.loading = false;
         }
-      })
+      });
     }
   }
